+++ app/services/blueprints/base_service.rb
@@ -2,6 +2,5 @@
 
 module Blueprints
   class BaseService < ::BaseService
-    ValidationError = Class.new(StandardError)
   end
 end
